Create class to auto-adjust image levels
Create a class that will progressively scan an image, compute a binned
histogram, and use it to produce a guess at appropriate values for
automatic level adjustment. The class accepts parameters to control what
percent of pixels should be considered "outliers" and potentially
clipped, and what ratio of non-outlier to outlier range to accept. The
latter allows the range to be expanded to include "near outliers" so
that clipping can be avoided if the outliers are only slightly outside
the range that covers most of the image.

The progressive scan approach works as follows. First, determine the
square with power-of-two sides that fully encapsulates the image. Scan
the top left pixel of this square. Divide the square into four equal
parts and scan the top left pixels of the three squares besides the top
left. Repeat until the size of the squares is one pixel, at which point
all pixels will have been scanned. This allows the algorithm to produce
an initial (approximate) result quickly, which is then refined until the
entire image has been processed, using a mostly-uniform sampling with
extremely simple recursion and near-zero bookkeeping.

This is intended primarily for e.g. 16-bit images which frequently do
not use the full range of possible values and may have inconsistent
normalization.