Simplify management of OpenGL resources
Enable global context sharing, which allows us to be more conservative
about recreating OpenGL resources. Convert said resources to regular
member objects rather than smart pointers, as the extra level of
indirection is now unnecessary.

Note that we still need "clean-up" logic for our textures, as
QOpenGLTexture retains a pointer to the QOpenGLFunctions instance of the
context with which it was created, which leads to a crash in Qt code
trying to reuse the QOpenGLTexture after a context change, even though
the underlying OpenGL resource is fine. (This seems like a bug...)