Add function to load KWIVER image to OpenGL
Add a new function to directly transfer a kwiver::vital::image to an
OpenGL texture. Unlike the QImage path, which may need to transcode the
image one or more times (QOpenGLTexture's support for using QImage
directly requires that the QImage is in RGBA8888 format), this tries
fairly hard to upload the image directly, using swizzling to support
images with fewer than 4 channels (in particular, single channel images
are directly supported) and/or with reversed channel order. Even
plane-packed images are supported using texture arrays, although
properly using these requires a modified fragment shader to combine
planes. All of this should allow us to move imagery to the GPU faster.
Equally important, this also adds support for 16-bit (integer) and
32-bit (floating point) images.

What's missing:
- Unit tests
- Support for bitmap images
- Support for images that can't be directly uploaded

The latter two will likely be implemented when and if they are needed;
we don't often usually use bitmap images for rendering purposes, and
image data layouts that can't be directly uploaded should be rare.