Add unit test for image to texture conversion
Add the missing unit test for loading a KWIVER image into an OpenGL
texture. This is a somewhat rudimentary unit test in that it only covers
uint8 images with a small handful of layouts, but it should cover much
of the critical functionality, including plane-packed images and images
with reversed channels (negative channel-stride). Also, the code for
other value types is only very slightly different, but testing it is
annoyingly complicated. More test cases may be added in the future, as
needed.

(This actually caught at least one minor bug in the original
implementation, which has since been fixed and squashed into the
preceding commit.)