Improve VideoSource API
Modify VideoSource to more explicitly work with a TimeMap for seeking,
and to provide the complete TimeMap upon request rather than a QSet
(which, as an unordered collection, is somewhat likely to be unhelpful
on its own). Also modify its signals to consolidate the cases where data
is available and where data is not available, as we were not effectively
making use of the latter as distinct from passing a null pointer via the
former. Also include the time stamp in these signals, which will be
useful for users performing newly-possible inexact seeks.

This should help with certain use cases where sequential rather than
random access is desired.

Right now, this fixes users, but breaks KwiverVideoSource, which will be
ported to the new API in a subsequent commit.