Add "metadata" to video sources
Modify VideoSource to provide a new VideoMetaData instance along with
the available image, rather than just a time stamp. This new class
includes the time stamp, and will also include other information that is
useful to the GUI (for now, also the image name).

We are using our own class here in order to also include the time stamp,
which has providence separate from the metadata provided by the KWIVER
video_input, and also because there is a bit of cognitive dissonance
between the GUI's notion of metadata, which is generally frame-specific,
and the KWIVER notion where metadata is a separate data stream
independent of the frame images. However, in the future we may modify
our class to also provide the raw KWIVER metadata.