Require user-provided homography to be 4x4
Modify sealtk::gui::Player to only support user-provided homographies
that are QMatrix4x4. It turns out this is actually likely to be more
convenient to users, as QMatrix4x4 has some support for interoperability
with QTransform, and it also simplifies our internal logic. (If needed,
the QMatrix3x3 version can be readded later.)