Clean up various stuff in NOAA player
Modify sealtk::noaa::gui::Player to use raw pointers to track its
context menu actions; smart pointers are rarely needed for QObject
types, since they are normally lifetime-managed by being parented.
Simplify an instance of a signal that could be directly connected to
another signal. Modify other signal/slot connections where the slot is a
lambda to conform with code style, and to prefer assigning a parent to
the slot.