Don't use QString::{to,from}StdString
Avoid use of aforementioned methods, which assume that the std::string
is / should be UTF-8. While this is almost certainly correct on Linux,
it is almost as surely *wrong* on Windows. Instead, use the qtString and
stdString functions from qtStlUtil which use QString::{to,from}Local8Bit
under the hood.