Implement support for a "static" homography
Modify noaa::gui::Player to have a notion of a "master" window and
"slave" windows, with the latter able to use a transformation to map
their local space to the "master"'s space. Implement logic to allow the
user to load such a transformation.

This uses a vital::transform under the hood, which can include
non-linear transformations. Eventually we want to handle those better;
for now, we approximate by computing a homography matrix by mapping our
image corner points.

This is a "static" transformation in that the user loads a single
transformation for the entire data set (as opposed to having a separate
transformation per frame). Eventually we will probably want to add
support for per-frame transformations as well, which will likely rely on
the video source to provide the transformation.