"Clean up" imageToTexture test
Improve comments in some "test the test" code that was intentionally
left in a prior commit. The new comment explains the purpose of the
code, which is to help debug the test in case it is failing. Although
the test is passing right now, this code was left as it is likely to be
helpful if something should go wrong with the test in the future.