Remove C++14 requirement
Create a new "util" library providing our own implementation of
make_unique so that we can remove the C++14 requirement (without having
to resort to the obnoxious syntax for creating a std::unique_ptr that is
otherwise required). Modify CMake utilities to support INTERFACE
libraries. Drop CMAKE_CXX_STANDARD to 11 and make it REQUIRED.

This is not actually sufficient, as it depends on another pending
change which removes some use of std::unique_ptr.