Hook up level control in Player
Add new "contrast mode" options to gui::Player. The options are Manual
(user specifies minimum and maximum values) and Percentile (uses
AutoLevelsTask to compute values). This also includes the logic to keep
track of computed percentile levels per time-keyed image, and to fire
off new tasks as needed.

Modify NOAA GUI to use Percentile mode for the IR view. (For now, this
is hard-coded, as are the tolerance and deviation values.)