Introduce new VideoSource
Rewrite VideoSource to use a request-response model and to run the video
source in a separate thread. This will make it possible for a single
video source to service multiple consumers, which was not practical with
the previous design. It should also improve UI responsiveness if the
source is slow. As much of the thread management code as seems practical
has been implemented in the base class to ease the burden on
implementations.

Note that this is ONLY the new API. Accordingly, the build is horribly
broken as of this commit, as none of the users nor implementations have
been updated yet.