Update core library to use new video API
Make the necessary changes to other classes in the core library to use
the new video API, as introduced in the previous two commits. This still
doesn't get the code base as a whole building, but the core library now
builds and its updated unit tests pass.

Note that there is one semantic change "hiding" in here; a video source
factory is no longer associated with a video controller. Users need to
manually add created sources to their video controller. This is
necessary because the controller needs a requestor to be able to add the
source, and requestors (and controllers, for that matter) are unique to
consumers, while factories (and sources) are not. This changes is
therefore necessary for our overall objective in this branch, which is
for a video source to support multiple consumers.