Refactor frame distribution with controller
Refactor how consumers receive frames when using a VideoController.
Originally, I wanted to have the consumer provide a VideoRequestor when
registering the source with the controller, but this turns out to be
awkward to implement in the current code, and would preclude multiple
players showing the same video with the same controller. (I'm not sure
what the use case for that would be, but...)

Instead, introduce a new proxy class VideoDistributor which encapsulates
a VideoRequestor (and is owned by the VideoController) which translates
incoming frames into signals, like we had before. The difference is
that, rather than listening for a signal directly from the source,
consumers should obtain the distributor from the controller (which can
be done when the source is added) and listen to signals from the
distributor.