Refactor time handling in VideoController
Refactor how we handle the set of available times to use a QMap (more
specifically, a TimeMap<std::nullptr_t>, which is about as close as we
can get to a value-less QMap), as having an ordered set is often much
more convenient. Add TimeMap::keyMap method to conveniently produce
these. Modify VideoController to compute the set of available time
points on demand (and to mark it dirty when it changes) rather than at
every call. Make additional changes to delay the initial 'seek to start'
until after times are available.

This fixes a few UI bugs that were introduced by the asynchronous nature
of the new video API.