Fix some minor bugs
Add code to handle when the video source is unable to obtain an image.
Fix gui::Player::setImage to always update; previously, trying to clear
the image would not force an update. Adjust seeking code to provide a
request identifier so that we always get back a response if one of the
sources has no imagery for the requested time.

Eventually we may need to actually generate incrementing request
identifiers, in order to properly handle "snapping" of the time scrubber
value in a potentially high-latency environment. For now, this is just
not working (the scrubber does not always "snap"). However, it may be
that the VideoController obviates the need for this functionality (which
was adapted from another application which originated much of the design
of the new video API). More research is required.