Add architectural documentation
Create some design (architectural) documentation to provide an overview
of the planned architecture going forward. This reflects code that is
yet to be written, as well as code that has been pushed but not yet
merged.

The hope is that these will both serve to help clarify design intent
going forward, and to assist in understanding the architecture once
these designs have been implemented.

Note that one of the intended topics has not yet been written. This
should get fixed eventually, but meanwhile I want to make the portions
that have been written available, rather than sitting on them forever.