Create base classes for item data
Add base classes for item data models, and representations of item data.
The latter is more or less imported wholesale from ViViA, while the
former adds some minor convenience features. The representation class
defines a mechanism by which the underlying model can provide data in a
representation-agnostic manner, and serves as a base class for
implementing representations.

Also, tweak source_check.py to skip the copyright check on "trivial" C++
sources (defined as a file consisting of a single comment). This avoids
adding a mostly-gratuitous copyright notice to DataModelTypes.cpp, which
exists only to trigger CMake's automoc, and whose entire contents
consists of a comment explaining this.