Fix crash when changing video source
Modify gui::Player to disassociate itself from its video source when the
source is destroyed. I suspect this has been broken, but wasn't noticed
because a) the application only ever destroys a video source when
replacing it with a new one, and b) prior to the new video API, the old
source was only destroyed after assigning the new source to the Player.
When the new API was introduced, the application logic also changed to
destroy the old source first, which would leave the Player with a
dangling reference.