Add DirectoryListing class
Create a new class to assist in searching a directory for files matching
one or more specified type (extension) filters. This class returns the
matching files as a mapping (QHash) of the matching file name without
the type to the file's absolute path. This will be used to generate the
list of embedded pipelines that the user may execute.

Co-authored-by: Kyle Edwards <kyle.edwards@kitware.com>