Support pipelines with "missing" video
Modify KwiverPipelineWorker to allow adding a null video source, which
serves as a placeholder for pipelines that expect particular "types" of
sources to be at specific "indices". (For example, the NOAA GUI always
has EO=0, IR=1, but we might want to run IR-only pipelines.) This also
simplifies things for users that have a fixed set of video sources, but
want to be able to run pipelines that only need a subset of those
sources.

Modify the NOAA GUI to take advantage of these changes, lifting the
restriction that both EO and IR videos are loaded, and also tweak the UI
so that running pipelines is disabled until a video source is loaded.