"Require" C++14
Modify CMake code to request full C++14. Apparently this does work with
GCC 4.8.x, even though that was some ways from full C++14 support,
whereas requesting C++11 plus the required additional features does not.
While there was a plan to fix this at one point, apparently it is not
going to happen, and this is the "sanctioned" way to use C++1y with GCC
4.8.