"Require" C++14
Modify CMake code to request full C++14. Apparently this does work with
GCC 4.8.x, even though that was some ways from full C++14 support, while
CMake (as of writing) does not recognize that GCC 4.8.x supports the
cxx_lambda_init_captures feature. (A patch for that is in the works, but
won't be available until at least the next release of CMake.)