Fix ImageUtils test on GCC 4.8
GCC 4.8 chokes on raw string literals passed as arguments to a macro,
which are used by the aforementioned unit test. Work around this by
redefining QStringLiteral → QString when using GCC 4.8, which results in
constructing a QString using the string literal constructor instead.
This is less efficient, but at least it compiles.

GCC 4.8 also insists on an unnecessary space between the quotes and
suffix of a user-defined literal operator. (Clang 3.1 does also, but
later compilers don't require it. It seems this was an early restriction
in UDL definitions that was subsequently relaxed.)

(While we're at it, also make the UDL's constexpr; they really ought to
be...)