Create initial track model
Create a data model for supplying tracks using KWIVER data structures.
For now, it only supplies a fairly minimal set of data which is not
sufficient for display in the graphic (video) views, but should be
sufficient for a simple list view (and possibly a timeline view, though
that is not in the short term plans). Additional functionality will be
added later.

This includes an "unshared pointer"; a class which wraps a KWIVER shared
pointer in a way that provides partial value semantics, i.e. making a
copy causes the underlying object to be cloned, rather than just taking
an additional reference. This will facilitate making a clone of the
model, which is not yet implemented, but will be needed eventually.

Also clean up some cruft from AbstractItemModel.cpp.