Use KwiverTrackSource in NOAA GUI
Update NOAA GUI to use the new KwiverTrackSource. Start implementing
logic to actually do something with the resulting model. Remove logic
that disables loading "detections" until a video is loaded, as this
restriction is no longer necessary.

(Note that, for NOAA, we plan to continue to use "detections" in the UI,
even though we actually use tracks internally. This is partly because
kwiver::detected_object doesn't really have enough information on its
own to be usable.)

Co-authored-by: Kyle Edwards <kyle.edwards@kitware.com>