Improve KwiverTrackSource test
Modify track test helper functions to optionally not rely on the index
of tracks within the model. This allows the KwiverTrackSource to be
robust against KWIVER returning the tracks in some order other than the
one we naïvely expect. (This seems to happen with some builds, for some
reason...)