Add ScalarFilterModel
Create ScalarFilterModel, which is the long-promised class which filters
a model based on lower and/or upper bounds on one or more data roles.
(Because it was fairly easy to implement, and is probably cheaper than
having separate proxy models per filter, ScalarFilterModel supports as
many filters as there are data roles. For obvious reasons, only one
lower and one upper bound per role is supported.)