Fix namespace metaobject visibility
Redeclare the metaobjects for our namespaces so that we can
export-decorate them... when we aren't on Windows. This is really ugly
and should be fixed on Qt's side, but for now this is required to not
get unresolved externals.

...And in fact, it has been fixed, but only in 5.14, which is some time
from release. (As of writing, even 5.13 is not out yet.) Accordingly,
also add the "nice" code, behind a version-check.