Work around resources not being exported
Add an RAII class to make the NOAA gui library's resources available,
following the pattern that already existed for the common gui library.
This allows us to work around not being able to directly call
Q_INIT_RESOURCE on the library's resources, since the required symbol is
not exported (and cannot be, as it comes from rcc-generated code).