Fix signed/unsigned issue
Refactor code slightly to avoid applying the unary negation operator to
an unsigned value. Thanks to two's-complement (and the fact that the
value is subsequently cast to signed), the code appears to have been
working anyway, but MSVC understandably warns about it. Moving the
negation out of the cast does not change the effective logic, but should
avoid the warning.