Update view transform on changing scene transform
Tweak our code to set the scene transform (homography) to also force an
update of the view transform. This is needed because the view transform
depends on whether or not we have a non-identity scene transform active.
Previously, the view would not update correctly when a scene transform
is initially loaded.