Add Meta-Information to SeekMode
Add meta-information to the SeekMode enumeration used by TimeMap. This
improves debugging of values of the same, and also allows the type to be
used in signals/slots and QVariant. (This will allow us to improve the
unit test for TimeMap by passing a SeekMode directly, rather than having
to cast to/from an integer type.)

This depends on https://github.com/Kitware/qtextensions/pull/28 in order
for CMake to correctly run AUTOMOC on this header.