Style fixes
Add a configuration file for running uncrustify (note: depends on latest
development version and some pending PR's). Fix various style issues as
discovered with the help of uncrustify.

Note that we are *not* just blindly applying uncrustify; it still has a
number of bugs that are showing up in our code base. Instead, these
changes are hand-picked from changes that uncrustify wants to make, and
in some cases, additional changes have been made (especially to include
orders) that are beyond uncrustify's current abilities, but which were
noticed as a result of changes that uncrustify did make.

Additionally, the configuration is still a work in progress and will
likely need some further tweaking.