Fix vertex indices in detection rendering
Tweak detection rendering code to account for vertices being tuples,
while our vertex data array is an array of scalars. This was causing us
to render too many vertices, and (sort of) at the wrong offsets. This
appears to have gone unnoticed as apparently OpenGL has built-in out of
bounds protection, and simply doesn't do anything when asked to render
indices that don't exist.