Refactor TimeStamp comparisons
Rewrite the comparison operators for TimeStamp to be less heavily
macro-dependent. There is still some use of macros, but the macro
definitions are much shorter, with more of the shared code moved to a
freestanding helper function.

This ends up being less terse, but ought to be more readable, and less
susceptible to butchering by style tools (which often have a hard time
with complex macros), and ought to generate the same code with decent
compilers when optimization is enabled.