Enhance track tests
Teach the KwiverTrackModel and KwiverTrackSource tests to also
understand (and in the former case, test) classifications. Also update
KwiverDetectionsSink test as needed for API changes.

Classification testing for KwiverDetectionsSink will be added in a
future commit. The KwiverTrackSource does not test classifications
because our available reader (kw18) does not support them.