Improve VideoSourceFactory API
Rename loadVideoSource method, which doesn't actually do anything but
tell the factory to emit a signal that *asks* for a source to be loaded,
to requestVideoSource. Replace FileVideoSourceFactory::loadFile method
with loadVideoSource method, which instead takes a URI and is moved to
the base class (VideoSourceFactory). A URI is more flexible and not
specific to a local file system, and should be capable of describing the
location of most video sources we would care to load.

Besides improved generality, this will also allow us to pass parameters
to a video source factory, including a FileVideoSourceFactory, by means
of URI queries.