Add UI to specify filters for image directory
Create a new class to allow the user to input a set of shell glob file
name filters. Use this to ask the user for the same when creating a
video source from a directory.

Eventually, this will provides an easier way for users to use only
select files within a directory as a video source (compared to manually
creating an image list file). However, the video source currently does
not support passing filters.