Render "shadow data"
Modify the Player to also use shadow data as a source for detections. If
a source is valid (has both a model and a transform), and the primary
source did not already provide a detection with the same ID, then
detections provided by shadow data sources will be rendered (with their
respective transform applied).