Add "supplemental" data to sinks
Modify AbstractDataSink to add a notion of "supplemental" data; that is,
data using a different coordinate space. This mirrors the addition (in a
different branch) of "shadow data" to the views, which provides a means
of displaying data from different modalities / coordinate spaces. This
will make it possible for data sinks to similarly support writing out
such data.