Refactor logic to merge tracks
Make some base Player slots virtual, so that the NOAA Player can
intercept the data being provided by the same. Leverage this to move
track merging from the base Player, where it very arguably doesn't
belong, considering that the action is owned by the NOAA player, and
instead move it to the NOAA player. Also leverage our new knowledge to
disable the action when we know it can't be done. Finally, tweak the
error messages slightly.