Refactor Sink test
Extract the utility model and VideoSource from the KwiverDetectionsSink
unit test so that they can be reused. Both of these will also be useful
for creating a unit test for the recently added KwiverTracksSink. (The
compareFiles would also be useful, but it's short and has a bunch of
dependencies, making it not really worth the effort.)