Fix centering on tracks
Modify the logic that tries to center on a detection to also match the
detection time against the request time. This didn't matter when we only
had detections, but now that we have tracks, it was causing the logic to
execute for every state in the track, often resulting in centering on
the wrong coordinates.