Remove gratuitous smart pointer
Don't use a std::unique_ptr to store the VideoController. This is a
QObject that is already managed by the QObject hierarchy, and will be
cleaned up via the same. (We are somewhat fortunate, in fact, that the
smart pointer was deleting it before the parent tries to reap children,
or we would be seeing a double free. While we can *probably* rely on
this, the smart pointer isn't necessary, and just using the QObject
hierarchy simplifies things slightly.)