Don't add empty vertex info
Tweak how we build the set of vertex information for detections to be
rendered to do a better job of avoiding creating empty hunks. These
wouldn't be rendered anyway, as the code that loops over the set of
vertices per box would be an empty loop, but this may allow us to skip
rendering entirely in more cases.