Create a representation to render detections
Create new DetectionRepresentation to encapsulate the code used to
render detections. This allows CreateDetectionPlayerTool to also use
this code directly without having to add special-purpose shims in the
Player, and also allows us to simplify the logic for determining what
color to use. Additionally, modify our shaders to take only one
transformation, which is precomputed on the CPU side, rather than
performing the same matrix multiplication for every vertex.

Besides the simplifications and improved encapsulation, this also allows
us to move the vertex buffer for pending detections to the tool, so that
we are not creating a new one every time we repaint.