Use QVector for detection vertex info
Switch back to using QVector, rather than std::vector, for the list of
detected object vertex information. This information is also needed for
picking, and since that is expected to involve asynchronous processing,
it will be nice to be able to avoid making a copy of the data. (If it
needs to be changed before the computation is complete, the master
copies are cleared first, so in theory we never need to make a copy, but
processing might hang on to a prior instance.)

For similar reasons, move the vertex data to a persistent array, rather
than one that gets dumped as soon as it is copied into the GPU buffer.