Don't render shadow data with no transform
Add a new virtual method to Player to test whether or not it has a
transform (the base implementation never does). Add a method to
noaa::Player to set the transform.

All of this is needed because we need to distinguish between a valid
identity transform and no transform, which was not possible from just
the homography.