Finish hooking up detection picking
Add remaining logic in the NOAA Window to respond to a Player signalling
that a track (detection) has been picked in the image view, so that
picking a detection will select the corresponding track(s) in the track
list.

Because this modifies the selection and not the current index, it does
not trigger 'focusing' on the track. This is intentional, as panning the
view when a detection is picked would most likely be distracting.