Allow panning with left mouse button
Modify Player to also allow panning with the left mouse button. Panning
with the middle button is still allowed (as this is the only way to pan
when creating a detection), and the Player will remain in panning mode
as long as either button is pressed.

Unfortunately, this interacts badly with creating detections, because
the tool needs to be able to preempt mouse events, and was not designed
to do so. This commit includes logic to permit such preemption, but the
tool will be fixed in a subsequent commit.