Overhaul CreateDetectionPlayerTool event handling
Modify how CreateDetectionPlayerTool handles mouse events so that it
appropriately preempts event handling by the Player, which is necessary
to prevent panning (with the left button) while creating an event. (See
also previous commit.)

Additionally, add the ability to create a detection by clicking opposite
corners, rather than clicking and dragging out a box, which has been
annoying me for some time. Also add ability to reset creation if the
image changes, which can happen in the new mode.

This still needs a cancellation mechanism. This will be implemented in a
subsequent commit.