Preserve original icon theme as fallback
Modify how we set up our own icon theme to also preserve the original
theme as the fallback theme, so that stuff like file dialogs doesn't end
up with screwy icons. Unfortunately, this relies on a fairly gratuitous
hack that involves running an external program, because it seems that
QIcon::themeName does not work.