Don't count items that are not visible
Modify ClassificationSummaryRepresentation to ignore items from the
source model which are not visible. I believe the original theory was
that, when we eventually have filtering on individual types, the 'best'
classification will take care of hiding things that do not pass filters.
However, right now, the non-type-specific filtering we have only
modifies the visibility.

(Note also that this branch was originally started before the score
filter was merged, and has subsequently been rebased, so we're also
cleaning up fallout from that.)