Fix centering on selected tracks
Fix an issue where centering on a selected track was not working in some
views when a homography is loaded, due to using the wrong effective
image size. Refactor determining the effective image size into a helper
method, as it is needed in multiple places. Fix some minor style issues
(not using `this->` to access a member variable) and remove some
unnecessary conversions by keeping math in double precision for longer,
rather than casting to float, promoting to double, and casting back
down.