Make NOAA GUI plugins configurable
An ongoing issue has been that we expect the NOAA GUI to be built to use
KWIVER plugins from an additional dependency (VIAME), but don't want
this dependency to be mandatory. As a result, we have been maintaining a
branch that changes some of the plugins used.

Instead, arrange for the ability to configure the relevant plugin
selection at configure time. This maintains our goal of minimizing
dependencies out-of-the-box, but will make it possible to punt plugin
selection to configure time, rather than having to maintain a branch to
change the defaults.