Add ability to update states in KwiverTrackModel
Add a new method to KwiverTrackModel that takes a track state to be
added to, or modified within, an existing track. This will make it
possible to add additional detections to an existing track without
having to create a new track and then merge the two tracks together.
(This is a little ugly because it requires figuring out twice where the
state will be inserted; once on our end in order to correctly signal
what changes will occur in the model, and again in KWIVER to modify the
underlying track object. Alas, there is no good way around this for
now.)