Don't check copyrights on QRC files
Modify the source check script to ignore QRC files when checking
copyrights. As these are effectively just lists of resource files to be
compiled in, there is little value in even asserting copyright (which
technically isn't necessary anyway), and these files are subject to
being rewritten by Designer, which strips any notice we try to add.