Don't 'remember' drag button on left click
Tweak our input handling so that a left click (without dragging) does
not leave the left button in our set of 'buttons active for dragging'.
This fixes an issue that can cause left click to select a detection to
not work. This also fixes another way in which creation can get 'stuck'
panning when it should not be doing so. (A different way of triggering
this was previously fixed by 6923dd3887c2.)