Sort video frames by time
Enable the option to tell video_input_image_list to sort the set of
frames according to their timestamps. This will alter their frame
numbers and may prevent some potential issues that could arise
particularly with pipelines because we generally process things in
temporal order and expect that to match frame order. (In particular,
there are some places that we expect things to arrive in temporal order
that try to append to KWIVER tracks, which will fail if the frame number
order does not match the temporal order.)

Note that this "depends" on Kitware/kwiver#1007, which adds the
aforementioned functionality to KWIVER. Things will still run without
that, but this change won't do anything.