Show both shortcuts on menu
Tweak the NOAA UI so that the menus show both shortcuts for the "main"
track-related actions. This is very much something of a hack as it
relies on an (AFAIK) undocumented feature where anything following a tab
character in an action's text overrides what is shown for the action's
shortcut. This also means that the text has to be kept in sync with the
actual shortcuts, but this seems worth it for the UX improvement.