Add ClassificationFilterModel
Create ClassificationFilterModel, a class which filters the
classification data of a model. This is related to ScalarFilterModel,
but unlike that class, which directly filters on one or more values
which can be of multiple roles, this class specifically filters the
individual entries from an item's ClassifiationRole data, with indirect
effects on ClassificationTypeRole and ClassificationScoreRole, as well
as (of course) VisibilityRole.