Add ClassificationFilterWidget
Create a widget to provide a set of filters to allow users to filter
based on an item's classification. Create a subclass of this which uses
the type names known to KWIVER's detected_object_type to automatically
populate the filters.

Note that this depends on https://github.com/Kitware/kwiver/pull/1010.