Ensure track types are sorted in the delegate
Modify TrackTypeDelegate to explicitly sort the collection of types we
get from KWIVER (using locale-aware sort, also). This will ensure that
the types are presented to the user in the best possible order,
regardless of whether what KWIVER gives us is in order or not (which may
or may not be changing upstream).