Allow changing text when no video is loaded
Add a method to Player to allow users (of the class, not necessarily of
an application) to change the text that is displayed when no video is
loaded. This is desirable as the default assumes that video will be
loaded via a context menu, which may not be the case for all
applications.